# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GymWithOccupancyBusynessLevel =
  (typeof GymWithOccupancyBusynessLevel)[keyof typeof GymWithOccupancyBusynessLevel];

export const GymWithOccupancyBusynessLevel = {
  quiet: "quiet",
  moderate: "moderate",
  busy: "busy",
  very_busy: "very_busy",
} as const;

export interface GymWithOccupancy {
  id: number;
  name: string;
  capacity: number;
  currentCount: number;
  busynessLevel: GymWithOccupancyBusynessLevel;
}

export interface CheckinBody {
  sessionId: string;
}

export interface CheckinResponse {
  success: boolean;
  /** @nullable */
  gymId: number | null;
  sessionId: string;
}

export interface SessionStatus {
  sessionId: string;
  /** @nullable */
  checkedInGymId: number | null;
}

export interface ErrorResponse {
  error: string;
}