# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: gyms
    description: Gym check-in operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /gyms:
    get:
      operationId: listGyms
      tags: [gyms]
      summary: List all gyms with current check-in count
      responses:
        "200":
          description: List of gyms with live check-in counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GymWithOccupancy"
  /gyms/session/{sessionId}:
    get:
      operationId: getSessionStatus
      tags: [gyms]
      summary: Get which gym a session is currently checked into
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session check-in status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionStatus"
  /gyms/{id}/checkin:
    post:
      operationId: checkInGym
      tags: [gyms]
      summary: Check in to a gym
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckinBody"
      responses:
        "200":
          description: Successfully checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinResponse"
        "409":
          description: Already checked in somewhere
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /gyms/{id}/checkout:
    post:
      operationId: checkOutGym
      tags: [gyms]
      summary: Check out of a gym
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckinBody"
      responses:
        "200":
          description: Successfully checked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinResponse"
        "404":
          description: Not currently checked in here
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    GymWithOccupancy:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        capacity:
          type: integer
        currentCount:
          type: integer
        busynessLevel:
          type: string
          enum: [quiet, moderate, busy, very_busy]
      required:
        - id
        - name
        - capacity
        - currentCount
        - busynessLevel
    CheckinBody:
      type: object
      properties:
        sessionId:
          type: string
      required:
        - sessionId
    CheckinResponse:
      type: object
      properties:
        success:
          type: boolean
        gymId:
          type: ["integer", "null"]
        sessionId:
          type: string
      required:
        - success
        - gymId
        - sessionId
    SessionStatus:
      type: object
      properties:
        sessionId:
          type: string
        checkedInGymId:
          type: ["integer", "null"]
      required:
        - sessionId
        - checkedInGymId
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error