# lib/api-zod/src/generated/types/gymWithOccupancy.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GymWithOccupancyBusynessLevel } from "./gymWithOccupancyBusynessLevel";

export interface GymWithOccupancy {
  id: number;
  name: string;
  capacity: number;
  currentCount: number;
  busynessLevel: GymWithOccupancyBusynessLevel;
}